Recurring oil leak check implemented
PyMOT will now scan a vehicle's MOT history for recurring oil leaks which have been documented in two or more MOT's. For example, if an oil leak is documented in MOT's from 2016 and 2017, this will be flagged for inspection.
Only repeating occurences are recorded. For example, oil leaks documented from 2011 and 2015 tests are *not* flagged.

 Future versions of this will ignore retests.